requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
pydantic>=2.6.0
groq>=0.9.0
python-dotenv>=1.0.0
supabase>=2.4.0
asyncio
